docs(logger): rework attributes doc sections
